// src/redux/reducers/user/user.type.ts
export type UserDoctor = {
    avatar: string
    email: string
    evaluate: string
    experience: string
    specialized: string
    username: string
    _id: string
    address: string
    phoneNumber: string
    role: string
    appointmentCount: number
    status: string
}

export type handleUpdatePermission = {
    _id: string
    role: string
    status: string
}